This is a To-do-list web apk
I made this app in 3 days. I use HTML, CSS, and JS to build this web app. The difference between my To-do-list app and others is I implement the date and time features of adding the task to the list. Which helps users remember when they add this task to the list.